Expose jsonMessage for file events
Extract message processing to simplify the design (and tests)
