+++ src/main/java/com/pubnub/api/models/consumer/pubsub/PNEvent.java
@@ -1,0 +1,4 @@
+package com.pubnub.api.models.consumer.pubsub;
+
+public interface PNEvent {
+}
