+++ src/main/java/com/pubnub/api/models/consumer/pubsub/files/PNFileEventResult.java
@@ -1,13 +1,15 @@
 package com.pubnub.api.models.consumer.pubsub.files;
 
+import com.google.gson.JsonElement;
 import com.pubnub.api.models.consumer.files.PNDownloadableFile;
+import com.pubnub.api.models.consumer.pubsub.PNEvent;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NonNull;
 
 @Data
 @Builder
-public class PNFileEventResult {
+public class PNFileEventResult implements PNEvent {
     @NonNull
     private final String channel;
     @NonNull
@@ -16,4 +18,5 @@ public class PNFileEventResult {
     private final Object message;
     @NonNull
     private final PNDownloadableFile file;
+    private final JsonElement jsonMessage;
 }
